Removed auto-removal of conda curl
As it is, this role only runs if the user specifies installation of curl, so having the removal of conda curl in by default makes no sense. Plus, if a user has explicitly specified that curl be installed with conda, they probably don't want it removed automatically, conflicts or no. As such, there is no circumstance in which this task is a good idea.